fix: empty line issue
sometimes during export whitespaces are not treated as part of plain
text are are simply left as not paresed. Fixed it by hardcoding a
whitespace in link. That should probably be fixed in post processing,
by parsing the whole exported text ensuring all the whitespaces are
part of block after